fixed bloom filter
Fixed some calculations and indexing for/of bloomfilter

+++ Backend/ServiceLayer_New/Services/IBloomFilterService.cs
@@ -2,7 +2,7 @@
 {
     public interface IBloomFilterService
     {
-        List<UInt64> GetHash(string username, long quantity, Double percentFP, int bitsRequired, int hashFuncRequired);
+        List<UInt64> GetHash(string username, long quantity, Double percentFP, long bitsRequired, int hashFuncRequired);
         byte[] MurmurHash128(byte[] receivedString);
         byte[] XxHash128(byte[] receivedString);
     }
